requirements.txt
anyio==3.6.2
asttokens==2.2.1
backcall==0.2.0
beautifulsoup4==4.11.1
black==22.12.0
certifi==2022.12.7
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
comm==0.1.2
cssselect==1.2.0
debugpy==1.6.4
decorator==5.1.1
entrypoints==0.4
executing==1.2.0
filelock==3.9.0
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
huggingface-hub==0.11.1
idna==3.4
ipykernel==6.19.4
ipython==8.8.0
jedi==0.18.2
jupyter_client==7.4.8
jupyter_core==5.1.2
lxml==4.9.2
markdownify==0.11.6
matplotlib-inline==0.1.6
mypy-extensions==0.4.3
nest-asyncio==1.5.6
notion-client==2.0.0
numpy==1.24.1
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
packaging==22.0
parso==0.8.3
pathspec==0.10.3
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==2.6.2
plumbum==1.8.1
ply==3.11
prompt-toolkit==3.0.36
protobuf==3.20.0
psutil==5.9.4
ptyprocess==0.7.0
pure-eval==0.2.2
Pygments==2.14.0
python-dateutil==2.8.2
python-dotenv==0.21.0
PyYAML==6.0
pyzmq==24.0.1
readability-lxml==0.8.1
regex==2022.10.31
requests==2.28.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
sentencepiece==0.1.97
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
stack-data==0.6.2
tokenizers==0.13.2
tomli==2.0.1
torch==1.13.1
tornado==6.2
tqdm==4.64.1
traitlets==5.8.0
transformers==4.25.1
tweepy==4.12.1
typing_extensions==4.4.0
urllib3==1.26.13
wcwidth==0.2.5
